Move dnd classes to dnd package
