+++ vaadin-grid-flow/src/main/java/com/vaadin/flow/component/grid/dnd/DropMode.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.grid;
+package com.vaadin.flow.component.grid.dnd;
 
 /**
  * Defines the locations within the Grid row where an element can be dropped.
@@ -63,7 +63,7 @@ public enum DropMode {
         this.clientName = clientName;
     }
 
-    String getClientName() {
+    public String getClientName() {
         return clientName;
     }
