+++ vaadin-grid-flow/src/main/java/com/vaadin/flow/component/grid/dnd/GridDragEndEvent.java
@@ -13,10 +13,11 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.grid;
+package com.vaadin.flow.component.grid.dnd;
 
 import com.vaadin.flow.component.ComponentEvent;
 import com.vaadin.flow.component.DomEvent;
+import com.vaadin.flow.component.grid.Grid;
 
 /**
  * Drop event on an HTML5 drop target {@link Grid} row.
