+++ vaadin-grid-flow/src/main/java/com/vaadin/flow/component/grid/dnd/GridDragStartEvent.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.flow.component.grid;
+package com.vaadin.flow.component.grid.dnd;
 
 import java.util.ArrayList;
 import java.util.Collections;
@@ -24,6 +24,7 @@ import com.vaadin.flow.component.ComponentUtil;
 import com.vaadin.flow.component.DomEvent;
 import com.vaadin.flow.component.EventData;
 import com.vaadin.flow.component.UI;
+import com.vaadin.flow.component.grid.Grid;
 
 import elemental.json.JsonArray;
 import elemental.json.JsonObject;
